Separate vertices for hard edge normals
The issue with shared vertices is that same vertex has normal vector
information for all faces that share that vertex. This makes sense
for smooth surfaces, but not for hard edges like cube has.

Depending on number of segments in prism or pyramid, they might serve
function of cylinder or cone with desire for smooth normals across ring
edges. This is not supported at the moment.