Quad shape from triangles, with subdivisions
A quad shape built from triangles, that can be subdivided as necessary
to model details. It can be bent into terrain or clothes.

For better performance, triangle strips should be used to reduce number
of vertices. The rest of lib doesn't use triangle strips, so this
optimization is not implemented.

For less artifacts in final model, it would be better to useequilateral
triangles in the mesh. Then the overal shape would have to be a triangle
or a hexagon, and most of the time a quad is more useful.

This mesh actually reuses vertices in adjecent triangles, which is not
how the rest of library works. Should add a flag to control this.

Also, a quad is a 2D shape and not a solid...